Refactor transaction UI imports and enhance component structure
- Update imports in Transaction, TransactionMenu, and related components to directly reference specific UI elements, improving clarity and maintainability.
- Remove the centralized index file for transaction UI components to streamline the import process and reduce complexity.
- Enhance TransactionMessage and TransactionToolbar components with additional props for better state handling and UI flexibility.
