+++ apps/app/src/app/@buyer/transaction/ui/Transaction.tsx
@@ -8,7 +8,8 @@ import { type FC, useRef, useState } from "react";
 import { useUpload } from "~/app/@common/gallery/hook/useUpload";
 import { ListingPrice } from "~/app/@common/listing/ui/ListingPrice";
 import { LocationBadge } from "~/app/@common/location/ui/LocationBadge";
-import { TransactionChat, TransactionMenuButton } from "~/app/@common/transaction/ui";
+import { TransactionChat } from "~/app/@common/transaction/ui/TransactionChat";
+import { TransactionMenuButton } from "~/app/@common/transaction/ui/TransactionMenuButton";
 import { TransactionEntryList } from "~/app/@common/transaction-entry/ui/TransactionEntryList";
 import { TransactionMenu } from "./TransactionMenu";
