+++ apps/app/src/app/@buyer/transaction/ui/status/OpenToolbar.tsx
@@ -1,6 +1,6 @@
 import type { tTransaction } from "@zbav-se.me/sdk/api/buyer";
 import type { FC } from "react";
-import { TransactionButtonUi } from "~/app/@common/transaction/ui";
+import { TransactionButtonUi } from "~/app/@common/transaction/ui/TransactionButtonUi";
 import type { TransactionMenuButton } from "~/app/@common/transaction/ui/TransactionMenuButton";
 import { CloseButton } from "../button/CloseButton";
 import { SuccessButton } from "../button/SuccessButton";
