--- apps/app/src/app/@common/transaction/ui/index.ts
@@ -1,6 +1,0 @@
-export { MessageButtonUi } from "./MessageButtonUi";
-export { TransactionButtonUi } from "./TransactionButtonUi";
-export { TransactionChat } from "./TransactionChat";
-export { TransactionMenuButton } from "./TransactionMenuButton";
-export { TransactionMessage } from "./TransactionMessage";
-export { TransactionToolbar } from "./TransactionToolbar";
