+++ apps/app/src/app/@seller/transaction/ui/status/DisputeToolbar.tsx
@@ -1,6 +1,6 @@
 import type { tTransaction } from "@zbav-se.me/sdk/api/seller";
 import type { FC } from "react";
-import { TransactionButtonUi } from "~/app/@common/transaction/ui";
+import { TransactionButtonUi } from "~/app/@common/transaction/ui/TransactionButtonUi";
 import type { TransactionMenuButton } from "~/app/@common/transaction/ui/TransactionMenuButton";
 import { ResolveButton } from "../button/ResolveButton";
